+++ src/main/java/com/mulesoft/connectors/internal/models/openai/providers/OpenAIModerationModelNameProvider.java
@@ -1,9 +1,9 @@
 /**
  * (c) 2003-2024 MuleSoft, Inc. The software in this package is published under the terms of the Commercial Free Software license V.1 a copy of which has been included with this distribution in the LICENSE.md file.
  */
-package com.mulesoft.connectors.internal.connection.mistralai;
+package com.mulesoft.connectors.internal.models.openai.providers;
 
-import com.mulesoft.connectors.internal.models.MistralModelName;
+import com.mulesoft.connectors.internal.models.openai.OpenAIModelName;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
@@ -11,11 +11,11 @@ import org.mule.runtime.extension.api.values.ValueProvider;
 import java.util.Arrays;
 import java.util.Set;
 
-public class MistralAITextGenerationModelNameProvider implements ValueProvider {
+public class OpenAIModerationModelNameProvider implements ValueProvider {
 
   @Override
   public Set<Value> resolve() {
-    return ValueBuilder.getValuesFor(Arrays.stream(MistralModelName.values())
-            .filter(MistralModelName::isTextGenerationSupport).map(String::valueOf));
+    return ValueBuilder.getValuesFor(Arrays.stream(OpenAIModelName.values())
+            .filter(OpenAIModelName::isModerationSupport).map(String::valueOf));
   }
 }
