+++ src/main/java/com/mulesoft/connectors/internal/models/provider/ModelNameProvider.java
@@ -3,12 +3,9 @@
  */
 package com.mulesoft.connectors.internal.models.provider;
 
-import java.util.Arrays;
 import java.util.Set;
-import java.util.stream.Stream;
 
 import com.mulesoft.connectors.internal.models.ModelType;
-import com.mulesoft.connectors.internal.models.OpenAIModelName;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
